+++ CHANGELOG.md
@@ -8,9 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ---
 
 ## [Unreleased]
-- Added CC BY 4.0 license
+
 
 ---
+## [0.1.1] - 2025-07-02
+
+### Added
+- CC BY 4.0 license
+- Txt and toml export to compiler.py
+
+### Fixed
+- scindex code item id for Anomaly & Threat Detection
+- typo in changlog
 
 ## [0.1.0] - 2025-07-02
 
@@ -20,5 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `compiler.py` script for exporting the index to JSON and TOML.
 - GitHub repository structure including issue templates and `CONTRIBUTING.md`.
 
-[unreleased]: https://github.com/scindex/scindex/compare/v0.1.0...HEAD
-[0.0.1]: https://github.com/scindex/scindex/releases/tag/v0.1.0
+[unreleased]: https://github.com/scindex/scindex/compare/v0.1.1...HEAD
+[0.1.0]: https://github.com/scindex/scindex/releases/tag/v0.1.1
+[0.1.0]: https://github.com/scindex/scindex/releases/tag/v0.1.0